==> 00_config.ipynb <==
## Configuração do Pipeline de Experimentos

### Esse notebook define todas as configurações centralizadas do pipe, incluind:

- Configurações de S3 e pastas
- Validações do input
- Export pra JSON

# Parâmetros do pipeline"
BUCKET_NAME = "experimento-lucas-barbosa"
PASTA_RAW = "raw/"
PASTA_PROCESSED = "processed/"
PASTA_ARCHIVE = "archive/"
PASTA_ERROR = "raw/error/"  # Dentro de raw/
COLUNAS_OBRIGATORIAS = [
    'id_experimento',
    'id_customer',
    'grupo',
    'uso',
    'valor',
    'timestamp'
]
GRUPO_CONTROLE = "GC"

S3_RAW_PATH      = f"s3://{BUCKET_NAME}/{PASTA_RAW}"
S3_PROCESSED_PATH= f"s3://{BUCKET_NAME}/{PASTA_PROCESSED}"
S3_ARCHIVE_PATH  = f"s3://{BUCKET_NAME}/{PASTA_ARCHIVE}"

print("Configurações carregadas:")
print(f"  Raw: {S3_RAW_PATH}")
print(f"  Processed: {S3_PROCESSED_PATH}")
print(f"  Archive: {S3_ARCHIVE_PATH}")

#METRICAS_CONFIG = {
#    'taxa_conversao': 'Percentual de clientes que usaram',
#    'uso_medio':     'Média de uso por cliente',
#    'valor_medio':   'Valor médio por cliente',
#    'ticket_medio':  'Ticket médio (apenas quem gastou)'
#}

# Exportar JSON de configs
import json
import boto3

config = {
    'bucket_name': BUCKET_NAME,
    'pasta_raw': PASTA_RAW,
    'pasta_processed': PASTA_PROCESSED,
    'pasta_archive': PASTA_ARCHIVE,
    'colunas_obrigatorias': COLUNAS_OBRIGATORIAS,
    'grupo_controle': GRUPO_CONTROLE,
    #'metricas_config': METRICAS_CONFIG
}

# Salvar localmente
with open('config.json', 'w') as f:
    json.dump(config, f, indent=2)
print("✓ Configurações salvas localmente em 'config.json'")

# Fazer upload para S3 (sem try/except)
s3_client = boto3.client('s3')
s3_client.upload_file('config.json', BUCKET_NAME, 'config/config.json')
print(f"✓ Configurações enviadas para s3://{BUCKET_NAME}/config/config.json")
